baby steps towards setting up a boxen for ambient displays
* installs firefox
* creates the ambient_mozilla module and runs it for all installs
* starts firefox::aurora with profile included in the module

lots still to do:

* install, run, and vendor libs appropriately
* make the exec to start firefox depend on having aurora installed
* customize the profile
* ensure this is a stable way store the profile[1]
* turn on error reporting
* create and install cron.d to restart the browser nightly
* create and install cron.d to run `boxen` nightly
* enable a permanent do not disturb mode [2]
* softwareupdate -l ??

[1] storing the profile in boxen may be a silly idea. if anything in
it changes it will register as a dirty tree and fail any subsequent
updates. If that's the case we will need a dropbox or similar.

[2] DND resets itself periodically (24 hrs?) to prevent a user from
missing security fixes. Let's see if we can get an exec to turn it off
every night when we run boxen.