manage the user profile with the profile manager
Aurora no longer describes or honors the `-p` flag for specifying a path to a
profile. This patch uses the profile manager to create a new profile iff one
does not already exist. User.js is dropped into that profile and the process supervisor runs it via the profile manager.